+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Products.MeetingLiege Changelog
 ===============================
 
-4.2.7 (unreleased)
+4.2.7 (2022-09-22)
 ------------------
 
 - Adapted `MeetingItem._annex_decision_addable_states_after_validation`